+++ src/DataBlobs/services/HelpersDataWrapEntryBlob.jl
@@ -6,7 +6,7 @@
 # NOTE this is the convenience wrappers for entry and blob.
 
 """
-Get the data entry and blob for the specified blobstore or dfg retured as a tuple.
+Get the blob entry and blob for the specified blobstore or dfg retured as a tuple.
 Related
 [`getBlobEntry`](@ref)
 
@@ -42,6 +42,38 @@ $(METHODLIST)
 function deleteData! end
 
 
+# construction helper from existing BlobEntry for user overriding via kwargs
+BlobEntry(
+    entry::BlobEntry;
+    id::Union{UUID,Nothing} = entry.id, 
+    blobId::Union{UUID,Nothing} = entry.blobId, 
+    originId::UUID = entry.originId, 
+    label::Symbol = entry.label, 
+    blobstore::Symbol = entry.blobstore, 
+    hash::String = entry.hash,
+    origin::String = entry.origin,
+    description::String = entry.description, 
+    mimeType::String = entry.mimeType, 
+    metadata::String = entry.metadata, 
+    timestamp::ZonedDateTime = entry.timestamp, 
+    _type::String = entry._type, 
+    _version::String = entry._version,
+) = BlobEntry(;
+    id,
+    blobId,
+    originId,
+    label,
+    blobstore,
+    hash,
+    origin,
+    description,
+    mimeType,
+    metadata,
+    timestamp,
+    _type,
+    _version
+)
+
 function getData(
     dfg::AbstractDFG, 
     vlabel::Symbol, 
@@ -59,13 +91,12 @@ function getData(
     return de=>db
 end
 
-
-#TODO from blobstores
+# This is the normal one
 function getData(
     dfg::AbstractDFG, 
     blobstore::AbstractBlobStore, 
     label::Symbol, 
-    key::Union{Symbol,UUID, <:AbstractString, Regex}; 
+    key::Symbol; 
     hashfunction = sha256,
     checkhash::Bool=true
 )
@@ -82,82 +113,78 @@ function addData!(
     entry::BlobEntry, 
     blob::Vector{UInt8}; 
     hashfunction = sha256,
-    checkhash::Bool=true
+    checkhash::Bool=false
 )
     checkhash && assertHash(entry, blob, hashfunction=hashfunction)
-    de = addBlobEntry!(dfg, label, entry)
-    db = addBlob!(dfg, de, blob)
-    return de=>db
+    blobId = addBlob!(dfg, entry, blob) |> UUID
+    newEntry = BlobEntry(entry; blobId) #, size=length(blob))
+    addBlobEntry!(dfg, label, newEntry)
 end
-
-function addData!(dfg::AbstractDFG, blobstore::AbstractBlobStore, label::Symbol, entry::BlobEntry, blob::Vector{UInt8}; hashfunction = sha256)
-    assertHash(entry, blob; hashfunction)
-    de = addBlobEntry!(dfg, label, entry)
-    db = addBlob!(blobstore, de, blob)
-    return de=>db
-end
-
-#TODO check this one
-function addData!(
-    ::Type{<:BlobEntry}, 
-    dfg::AbstractDFG, 
-    label::Symbol, 
-    key::Symbol, 
-    blob::AbstractVector{UInt8}, 
-    timestamp=now(localzone());
-    id::UUID = uuid4(), 
-    hashfunction::Function = sha256
-)
-    fde = BlobEntry(key, id, timestamp, blob)
-    de = addBlobEntry!(dfg, label, fde)
-    return de=>blob
-end
-
-addData!(
-    dfg::AbstractDFG, 
-    blobstorekey::Symbol, 
-    label::Symbol, 
-    key::Symbol, 
-    blob::Vector{UInt8},
-    timestamp=now(localzone()); 
-    kwargs...
-) = addData!(
-        dfg,
-        getBlobStore(dfg, blobstorekey),
-        label,
-        key,
-        blob,
-        timestamp;
-        kwargs...
-    )
 
 function addData!(
     dfg::AbstractDFG, 
     blobstore::AbstractBlobStore, 
     label::Symbol, 
-    key::Symbol,
+    entry::BlobEntry, 
+    blob::Vector{UInt8}; 
+    hashfunction = sha256,
+    checkhash::Bool=false,
+)
+    checkhash && assertHash(entry, blob; hashfunction)
+    blobId = addBlob!(blobstore, entry, blob) |> UUID
+    newEntry = BlobEntry(entry; blobId) #, size=length(blob))
+    addBlobEntry!(dfg, label, newEntry)
+end
+
+
+addData!(
+    dfg::AbstractDFG, 
+    blobstorekey::Symbol, 
+    vLbl::Symbol, 
+    bLbl::Symbol, 
+    blob::Vector{UInt8},
+    timestamp=now(localzone()); 
+    kwargs...
+) = addData!(
+    dfg,
+    getBlobStore(dfg, blobstorekey),
+    vLbl,
+    bLbl,
+    blob,
+    timestamp;
+    kwargs...
+)
+
+#FIXME id used wrong
+function addData!(
+    dfg::AbstractDFG, 
+    blobstore::AbstractBlobStore, 
+    vLbl::Symbol, 
+    bLbl::Symbol,
     blob::Vector{UInt8}, 
     timestamp=now(localzone()); 
-    description="", 
-    mimeType = "application/octet-stream", 
-    id::UUID = uuid4(), 
+    description="",
+    metadata = "",
+    mimeType::String = "application/octet-stream", 
+    id::Union{UUID,Nothing} = nothing, #only assign if blobstore issued you an id
+    originId::UUID = uuid4(),
     hashfunction = sha256
 )
     #
-    @warn "ID's and origin IDs should be reconciled here."
-    entry = BlobEntry(
-        id = id, 
-        originId = id,
-        label = key, 
+    @warn "ID's and origin IDs should be reconciled here in DFG.addData!." maxlog=50
+    entry = BlobEntry(;
+        id, 
+        originId,
+        label = bLbl, 
         blobstore = blobstore.key, 
-        hash = bytes2hex(hashfunction(blob)),
-        origin = buildSourceString(dfg, label),
-        description = description, 
-        mimeType = mimeType, 
-        metadata = "", 
-        timestamp = timestamp)
+        hash = string(bytes2hex(hashfunction(blob))),
+        origin = buildSourceString(dfg, vLbl),
+        description, 
+        mimeType, 
+        metadata, 
+        timestamp)
 
-    addData!(dfg, blobstore, label, entry, blob; hashfunction)
+    addData!(dfg, blobstore, vLbl, entry, blob; hashfunction)
 end
 
 
@@ -169,17 +196,31 @@ function updateData!(
     hashfunction = sha256,
     checkhash::Bool=true
 )
-    checkhash && assertHash(entry, blob, hashfunction=hashfunction)
+    checkhash && assertHash(entry, blob; hashfunction)
+    # order of ops with unknown new blobId not tested
     de = updateBlobEntry!(dfg, label, entry)
     db = updateBlob!(dfg, de, blob)
     return de=>db
 end
 
-function updateData!(dfg::AbstractDFG, blobstore::AbstractBlobStore, label::Symbol,  entry::BlobEntry, blob::Vector{UInt8}; hashfunction = sha256)
+
+function updateData!(
+    dfg::AbstractDFG, 
+    blobstore::AbstractBlobStore, 
+    label::Symbol, 
+    entry::BlobEntry, 
+    blob::Vector{UInt8}; 
+    hashfunction = sha256
+)
     # Recalculate the hash - NOTE Assuming that this is going to be a BlobEntry. TBD.
-    newEntry = BlobEntry(entry.id, entry.blobId, entry.originId, entry.label, blobstore.key, bytes2hex(hashfunction(blob)),
-        buildSourceString(dfg, label),
-        entry.description, entry.mimeType, entry.metadata, entry.timestamp, entry._type, string(_getDFGVersion()))
+    # order of operations with unknown new blobId not tested
+    newEntry = BlobEntry(
+        entry; # and kwargs to override new values
+        blobstore = blobstore.key, 
+        hash = string(bytes2hex(hashfunction(blob))),
+        origin = buildSourceString(dfg, label),
+        _version = string(_getDFGVersion()),
+    )
 
     de = updateBlobEntry!(dfg, label, newEntry)
     db = updateBlob!(blobstore, de, blob)
@@ -188,20 +229,36 @@ end
 
 function deleteData!(
     dfg::AbstractDFG, 
-    label::Symbol, 
-    key::Symbol
+    vLbl::Symbol, 
+    bLbl::Symbol
 )
-    de = deleteBlobEntry!(dfg, label, key)
+    de = deleteBlobEntry!(dfg, vLbl, bLbl)
     db = deleteBlob!(dfg, de)
     return de=>db
 end
 
 
-deleteData!(dfg::AbstractDFG, blobstore::AbstractBlobStore, label::Symbol, entry::BlobEntry) =
-            deleteBlob!(dfg, blobstore, label, entry.label)
+function deleteData!(
+    dfg::AbstractDFG, 
+    blobstore::AbstractBlobStore, 
+    vLbl::Symbol, 
+    entry::BlobEntry
+) 
+    return deleteData!(
+        dfg, 
+        blobstore, 
+        vLbl, 
+        entry.label
+    )
+end
 
-function deleteData!(dfg::AbstractDFG, blobstore::AbstractBlobStore, label::Symbol, key::Symbol)
-    de = deleteBlobEntry!(dfg, label, key)
+function deleteData!(
+    dfg::AbstractDFG, 
+    blobstore::AbstractBlobStore, 
+    vLbl::Symbol, 
+    bLbl::Symbol
+)
+    de = deleteBlobEntry!(dfg, vLbl, bLbl)
     db = deleteBlob!(blobstore, de)
     return de=>db
 end