+++ src/Deprecated.jl
@@ -2,7 +2,23 @@
 ## ================================================================================
 ## Remove in v0.21
 ##=================================================================================
+@deprecate packFactor(dfg::AbstractDFG, f::DFGFactor) packFactor(f::DFGFactor)
 
+# #TODO check this one
+# function addData!(
+#     ::Type{<:BlobEntry}, 
+#     dfg::AbstractDFG, 
+#     vLbl::Symbol, 
+#     bLbl::Symbol, 
+#     blob::AbstractVector{UInt8}, 
+#     timestamp=now(localzone());
+#     id::UUID = uuid4(), 
+#     hashfunction::Function = sha256
+# )
+#     fde = BlobEntry(bLbl, id, timestamp, blob)
+#     de = addBlobEntry!(dfg, vLbl, fde)
+#     return de=>blob
+# end
 
 """
 $(TYPEDEF)