+++ src/DistributedFactorGraphs.jl
@@ -30,6 +30,9 @@ using UUIDs
 using Pkg
 using TensorCast
 using ProgressMeter
+using SHA
+
+using CSV
 
 # used for @defVariable
 import ManifoldsBase
@@ -174,6 +177,7 @@ export getPPEDict, getVariablePPEDict, getVariablePPE
 
 # CRUD & SET
 export getPPE,
+       getPPEs,
        getVariablePPE,
        addPPE!,
        updatePPE!,
@@ -201,6 +205,7 @@ export hasBlobEntry,getBlobEntry,addBlobEntry!,updateBlobEntry!,deleteBlobEntry!
 export incrDataLabelSuffix
 
 export getBlobEntries, listDataEntries, hasDataEntry, hasDataEntry
+export getBlobEntriesVariables
 export listDataEntrySequence
 # convenience wrappers
 export mergeDataEntries!
@@ -282,6 +287,18 @@ export
 ## CustomPrinting.jl
 export printFactor, printVariable, printNode
 
+# Data Blobs
+export InMemoryBlobStore
+export FolderStore
+export BlobEntry
+export getBlob, addBlob!, updateBlob!, deleteBlob!, listBlobEntries
+export listBlobs
+export BlobEntry
+# export copyStore
+export getId, getHash, getTimestamp
+# convenience wrappers
+export getData, addData!, updateData!, deleteData!
+
 ##==============================================================================
 ## Files Includes
 ##==============================================================================
@@ -291,7 +308,8 @@ export printFactor, printVariable, printNode
 include("entities/AbstractDFG.jl")
 
 # Data Blob extensions
-include("DataBlobs/DataBlobs.jl")
+include("DataBlobs/entities/BlobEntry.jl")
+include("DataBlobs/entities/BlobStores.jl")
 
 include("entities/DFGFactor.jl")
 
@@ -316,6 +334,10 @@ include("services/DFGVariable.jl")
 include("services/DFGFactor.jl")
 include("Deprecated.jl")
 include("services/CompareUtils.jl")
+#Blobs
+include("DataBlobs/services/BlobEntry.jl")
+include("DataBlobs/services/BlobStores.jl")
+include("DataBlobs/services/HelpersDataWrapEntryBlob.jl")
 
 # Include the FilesDFG API.
 include("FileDFG/FileDFG.jl")