+++ src/entities/DFGVariable.jl
@@ -94,7 +94,7 @@ VariableNodeData(variableType::InferenceVariable; kwargs...) = VariableNodeData{
 
 """
 $(TYPEDEF)
-Packed VariabeNodeData structure for serializing DFGVariables.
+Packed VariableNodeData structure for serializing DFGVariables.
 
   ---
 Fields:
@@ -181,26 +181,60 @@ getEstimateFields(::MeanMaxPPE) = [:suggested, :max, :mean]
 ## DFG Variables
 ##==============================================================================
 
-# Packed Variable
-Base.@kwdef struct PackedVariable
-    # NOTE: This has to match the order of the JSON deserializer as we're using OrderedStructs.
-    id::Union{UUID, Nothing}
+
+export Variable
+  
+"""
+    $(TYPEDEF)
+
+The Variable information packed in a way that accomdates multi-lang using json.
+"""
+Base.@kwdef struct Variable <: AbstractDFGVariable
+    id::Union{UUID, Nothing} = nothing
     label::Symbol
-    tags::Vector{Symbol}
-    timestamp::ZonedDateTime
-    nstime::Int
-    ppes::Vector{MeanMaxPPE}
-    blobEntries::Vector{BlobEntry}
+    tags::Vector{Symbol} = Symbol[]
+    timestamp::ZonedDateTime = now(tz"UTC")
+    nstime::String = "0"
+    ppes::Vector{MeanMaxPPE} = MeanMaxPPE[]
+    blobEntries::Vector{BlobEntry} = BlobEntry[]
     variableType::String
-    _version::String
-    metadata::String
-    solvable::Int
-    solverData::Vector{PackedVariableNodeData}
+    _version::String = string(_getDFGVersion())
+    metadata::String = "e30="
+    solvable::Int = 1
+    solverData::Vector{PackedVariableNodeData} = PackedVariableNodeData[]
 end
 
-StructTypes.StructType(::Type{PackedVariable}) = StructTypes.UnorderedStruct()
-StructTypes.idproperty(::Type{PackedVariable}) = :id
-StructTypes.omitempties(::Type{PackedVariable}) = (:id,)
+#IIF like contruction helper for packed variable
+function Variable(
+    label::Symbol,
+    variableType::String;
+    tags::Vector{Symbol} = Symbol[],
+    timestamp::ZonedDateTime = now(tz"UTC"),
+    solvable::Int = 1,
+    nanosecondtime::Int64 = 0,
+    smalldata::Dict{Symbol, SmallDataTypes} = Dict{Symbol, SmallDataTypes}(),
+    kwargs...
+)
+    union!(tags, [:VARIABLE])
+
+    pacvar = Variable(;
+        label,
+        variableType,
+        nstime = string(nanosecondtime),
+        solvable,
+        tags,
+        metadata = base64encode(JSON3.write(smalldata)),
+        timestamp,
+        kwargs...
+    )
+
+    return pacvar
+end
+const PackedVariable = Variable
+
+StructTypes.StructType(::Type{Variable}) = StructTypes.UnorderedStruct()
+StructTypes.idproperty(::Type{Variable}) = :id
+StructTypes.omitempties(::Type{Variable}) = (:id,)
 
 ##------------------------------------------------------------------------------
 ## DFGVariable lv2
@@ -331,6 +365,27 @@ Base.@kwdef struct DFGVariableSummary <: AbstractDFGVariable
     dataDict::Dict{Symbol, BlobEntry}
 end
 
+function DFGVariableSummary(
+    id,
+    label,
+    timestamp,
+    tags,
+    ::Nothing,
+    variableTypeName,
+    ::Nothing,
+)
+    return DFGVariableSummary(
+        id,
+        label,
+        timestamp,
+        tags,
+        Dict{Symbol, MeanMaxPPE}(),
+        variableTypeName,
+        Dict{Symbol, BlobEntry}(),
+    )
+end
+
+StructTypes.names(::Type{DFGVariableSummary}) = ((:variableTypeName, :variableType),)
 
 ##------------------------------------------------------------------------------
 ## SkeletonDFGVariable.jl
@@ -364,8 +419,8 @@ StructTypes.omitempties(::Type{SkeletonDFGVariable}) = (:id,)
 ##==============================================================================
 # Define variable levels
 ##==============================================================================
-const VariableDataLevel0 = Union{DFGVariable, DFGVariableSummary, SkeletonDFGVariable}
-const VariableDataLevel1 = Union{DFGVariable, DFGVariableSummary}
+const VariableDataLevel0 = Union{DFGVariable, DFGVariableSummary, Variable, SkeletonDFGVariable}
+const VariableDataLevel1 = Union{DFGVariable, DFGVariableSummary, Variable}
 const VariableDataLevel2 = Union{DFGVariable}
 
 ##==============================================================================